Add StrBuffer.java
Experiment 6:

StringBuffer Class Methods

Aim
Write a program in java for string handling which performs the following
i. Check the capacity of the StringBuffer object.
ii. Reverse the content of this string and convert the resultant string in upper case
iii. Read another string and append it to the resultant string of above.